addManageCourse
