+++ BusinessLayer/IService/ICourseService.cs
@@ -11,24 +11,17 @@ namespace BusinessLayer.IService
 {
     public interface ICourseService
     {
-        Task<CourseDto> CreateCourseAsync(CourseCreateDto courseCreateDto);
         Task<CourseDto> GetCourseByIdAsync(int courseId,string userId,string userRole);
         Task<CourseListDto> GetCourseByCourseId(int courseId);
-        Task<IEnumerable<CourseWithEnrollmentStatusDto>> GetAllCoursesAsync(string userRole, string userId);
-        Task<IEnumerable<CourseListDto>> GetAllCoursesForManagerAsync();
-        Task<IEnumerable<CourseListDto>> GetActiveCoursesAsync();
-        Task<IEnumerable<CourseListDto>> GetDraftCoursesAsync();
-        Task<IEnumerable<CourseListDto>> GetInactiveCoursesAsync();
-        Task<bool> CanApproveCourseAsync(int courseId);
-        Task UpdateCourseStatusAsync(int courseId, CourseStatus status);
+        Task<IEnumerable<CourseWithEnrollmentStatusDto>> GetAllCoursesAsync(string userId);
+        Task<IEnumerable<CourseListDto>> GetCoursesByStatusAsync(CourseStatus? status = null);
         Task<IEnumerable<CourseListDto>> GetCoursesInProgressAsync(string userId);
         Task<IEnumerable<CourseListDto>> GetCompletedCoursesAsync(string userId);
         Task<bool> IsUserEnrolledInCourseAsync(string userId, int courseId);
-        Task UpdateCourseAsync(int courseId, CourseUpdateDto courseUpdateDto);
-        Task DeleteCourseAsync(int courseId);
-        Task<IEnumerable<CourseWithEnrollmentStatusDto>> GetCoursesByTopicAsync(CourseTopic topic, string userRole, string userId);
 
-        Task<IEnumerable<CourseWithEnrollmentStatusDto>> SearchCourseAsync(string searchTerm, string userRole, string userId);
+        Task<IEnumerable<CourseWithEnrollmentStatusDto>> GetCoursesByTopicAsync(CourseTopic topic,   string userId);
+
+        Task<IEnumerable<CourseWithEnrollmentStatusDto>> SearchCourseAsync(string searchTerm,string userId);
 
         Task<CourseEnrollment> EnrollInCourseAsync(string userId, int courseId);
         Task<IEnumerable<CourseEnrollmentDto>> GetAllEnrollmentsForCourseAsync(int courseId);
@@ -44,5 +37,19 @@ namespace BusinessLayer.IService
         //getdetailCompleteCourse
         Task<CompletedCourseDetailDto> GetCompletedCourseDetailAsync(int courseId, string userId);
         Task<IEnumerable<CourseListDto>> GetCoursesWithoutSurveyOrInactiveSurveyAsync();
+
+        //Manager
+        Task<CourseDto> CreateCourseAsync(CourseCreateDto courseCreateDto);
+
+        Task<IEnumerable<CourseListDto>> GetAllCoursesForManagerAsync();
+        Task<bool> CanApproveCourseAsync(int courseId);
+        Task UpdateCourseStatusAsync(int courseId, CourseStatus status);
+        Task UpdateCourseAsync(int courseId, CourseUpdateDto courseUpdateDto);
+        Task DeleteCourseAsync(int courseId);
+        Task<IEnumerable<CourseListDto>> SearchCourseForManagerAsync(string searchTerm, CourseStatus? status = null);
+        Task<IEnumerable<CourseListDto>> GetCoursesByTopicForManagerAsync(CourseTopic topic, CourseStatus? status = null);
+
+
+
     }
 }
